Follow general elixir standards
When writing GenServer's it is typical to have a client interface section and a
server callbacks. It is important to note that the client interface runs in the
calling process and the server callbacks run in the server process. However,
this might make it a little harder to understand when first viewing the code
when you are first learning what a GenServer is. I'm wondering if we shouldn't
include this part but it's a little bit of a tricky decision.

Typically a GenServer that has only one process instance will be named with it's
module name, so instead of `:socket_broadcaster` as the name, this PR changes it
to `ChatApp.SocketBroadcaster` (note that they are both atoms).

Apply `mix format`

Use `Enum.reject` instead of `Enum.filter` with a `cond`

This PR also sets `reuseaddr` true which allows us to re-use ports that are
still in the `TIME_WAIT` tcp status. More info at:
https://superuser.com/questions/173535/what-are-close-wait-and-time-wait-states